yet another error in slicing of images

+++ pbt_gans/tflib_defs/save_images.py
@@ -31,7 +31,7 @@ def save_images(X, save_path):
         img = np.zeros((h*nh, w*nw))
 
     for n, x in enumerate(X):
-        j = n/nw
+        j = (int)n/nw
         i = n%nw
         img[j*h:j*h+h, i*w:i*w+w] = x
